Factor out sync cog to package.
