+++ bot/cogs/sync/syncers.py
@@ -1,10 +1,8 @@
 import logging
 from collections import namedtuple
-from typing import Callable, Iterable
 
-from discord import Guild, Role
+from discord import Guild
 from discord.ext.commands import Bot
-
 
 log = logging.getLogger(__name__)
 Role = namedtuple('Role', ('id', 'name', 'colour', 'permissions'))
@@ -56,31 +54,6 @@ async def sync_members(bot: Bot, guild: Guild):
     """
 
     current_members = await bot.api_client.get('bot/members')
+    site_members = {
+    }
 
-
-class Sync:
-    """Captures relevant events and sends them to the site."""
-
-    # The server to synchronize events on.
-    # Note that setting this wrongly will result in things getting deleted
-    # that possibly shouldn't be.
-    SYNC_SERVER_ID = 267624335836053506
-
-    # An iterable of callables that are called when the bot is ready.
-    ON_READY_SYNCERS: Iterable[Callable[[Bot, Guild], None]] = (
-        sync_roles,
-    )
-
-    def __init__(self, bot):
-        self.bot = bot
-
-    async def on_ready(self):
-        guild = self.bot.get_guild(self.SYNC_SERVER_ID)
-        if guild is not None:
-            for syncer in self.ON_READY_SYNCERS:
-                await syncer(self.bot, guild)
-
-
-def setup(bot):
-    bot.add_cog(Sync(bot))
-    log.info("Cog loaded: Sync")
